Revert "changed float to short"
This reverts commit 8544150a780f79a06f29982af2c858f06e9664d8.

+++ src/main/java/rinha/backend/MainHandler.java
@@ -36,7 +36,7 @@ public final class MainHandler {
   public void fraudScore(RoutingContext ctx) {
     FraudRequest request = FraudRequest.parse(ctx.body().asJsonObject());
 
-    short[] queryVector = vectorize(request);
+    float[] queryVector = vectorize(request);
 
     int fraudScore = knnSearch.search(queryVector);
